+++ pages/Library.py
@@ -11,8 +11,7 @@ parent_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 sys.path.insert(0, parent_dir)
 
 
-st.set_page_config(page_title="ASK Library", initial_sidebar_state="collapsed")
-
+    
 st.markdown( """ <style> [data-testid="collapsedControl"] { display: none } html, body, [class*="st-"] {font-family: "Source Sans Pro", "Arial", "Helvetica", sans-serif !important;}</style> """, unsafe_allow_html=True, )
 
 
@@ -43,6 +42,7 @@ def read_markdown_file(markdown_file):
 
 
 
+
 st.image("https://raw.githubusercontent.com/drew-wks/ASK/main/images/ASK_logotype_color.png?raw=true", use_container_width=True)
 
 back = st.button("< Back to App", type="primary")
@@ -64,13 +64,13 @@ with tab2:
     if df is not None:
         num_items = len(df)
         st.markdown("#### Library Overview")
-        st.markdown(f"ASK is loaded with **{num_items}** national documents (almost 9000 pages) including USCG Directives, CHDIRAUX documents and documents issued by the USCG Auxiliary National leadership. All these documents are located in public sections of the USCG and USCG Auxiliary websites (cgaux.org uscg.mil).  No secure content is included (i.e., content requiring Member Zone or CAC access. All documents are national. Regional requirements may vary, so check with your local AOR leadership for the final word. ")
+        st.markdown(f"ASK is loaded with **{num_items}** national documents (almost 9000 pages) including USCG Directives, CHDIRAUX documents and documents issued by the USCG Auxiliary National leadership. All these documents are located in public sections of the USCG and USCG Auxiliary websites (cgaux.org uscg.mil).  No secure content is included (i.e., content requiring Member Zone or CAC access). All documents are national in scope as of right now. Regional requirements may vary, so check with your local AOR leadership for the final word. ")
         st.markdown(f"{overview}")
         st.markdown("#### Library Catalog")
         st.markdown(f"{num_items} items. Last update: {last_update_date}")  
 
         # Display the DataFrame
-        display_df = df[['title', 'issue_date']]
+        display_df = df[['title', 'publication_number', 'organization', 'issue_date', 'expiration_date']]
         edited_df = st.data_editor(display_df, use_container_width=True, hide_index=False, disabled=True)
         isim = f'ASK_catalog_export{last_update_date}.csv'
         indir = edited_df.to_csv(index=False)
@@ -92,6 +92,7 @@ with tab4:
     roadmap = read_markdown_file("docs/roadmap.md")
     st.markdown(roadmap, unsafe_allow_html=True)
     
+    
 with tab5:
     st.markdown("#### Feedback")
     st.markdown("ASK's mission is to provide USCG Auxiliary members efficient, accuracete and easy access to the authoritative source of knowledge on any topic in the Auxiliary.")
@@ -104,3 +105,5 @@ with tab5:
             and a reference to the applicable policy (doc and page).''')  
                 
     st.markdown('Send an email to uscgaux.drew@wks.us.''')
+    
+